+++ Doc/Notes/Release/ChangeLog.md
@@ -12,6 +12,7 @@ Released on March XX, 2021
 - "Filtered" results are now visible after a `Clear()` operation (#78)
 - Can now see the progress bar during `Reload()`, so that I know the application is responding to my request. (#77)
 - Simplified Weevil's context menu. (#91)
+- A .NET callstack within a `IRecord` will be simplified to a more human readable form. (#88)
 
 #### Bug Fixes